refactor(agents): split grpc servicers from server module
