+++ FILEMAP.md
@@ -1,7 +1,7 @@
 # Repository File Map
 
 > Auto-generated by `scripts/generate_filemap.py` — 2026-04-20  
-> **269** tracked files
+> **270** tracked files
 
 ## Summary
 
@@ -9,7 +9,7 @@
 |---------------------|-------|
 | `docs/` | 55 |
 | `internal/` | 52 |
-| `agents/` | 44 |
+| `agents/` | 45 |
 | `tests/` | 35 |
 | `(root)` | 20 |
 | `scripts/` | 16 |
@@ -26,7 +26,7 @@
 
 | Extension | Count |
 |-----------|-------|
-| `.py` | 90 |
+| `.py` | 91 |
 | `.md` | 71 |
 | `.go` | 48 |
 | `.yaml` | 23 |
@@ -118,6 +118,7 @@
 │   ├── pyproject.toml
 │   ├── server.py
 │   ├── server_persona.py
+│   ├── server_servicers.py
 │   ├── sub_agents/
 │   │   ├── __init__.py
 │   │   ├── spawner.py
